Botei o service dentro da pasta de categories, e usei a url do environment.ts
