+++ src/ProFin.App/src/app/app.module.ts
@@ -11,9 +11,9 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ToastrModule } from 'ngx-toastr';
 import { CreateCategoryComponent } from './category/create-category/create-category.component';
-import { CategoryService } from './services/categories.service';
 import { ListCategoryComponent } from './category/list-category/list-category.component';
 import { EditCategoryComponent } from './category/edit-category/edit-category.component';
+import { CategoryService } from './category/services/categories.service';
 
 export const httpInterceptorProviders = [
   //{ provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
